Combine integration tests into single file
